restructure dirs, refactor modules, centralize todo
