--- data/franck-condon_legacy/harris_rkr_fc.csv
@@ -1,10 +1,0 @@
-3.547e-9 1.021e-7 1.430e-6 1.283e-5 8.264e-5 4.087e-4 1.608e-3 5.146e-3 1.362e-2 3.038e-2 5.765e-2 9.277e-2
-3.489e-8 9.090e-7 1.142e-5 9.071e-5 5.105e-4 2.167e-3 7.164e-3 1.875e-2 3.910e-2 6.537e-2 8.637e-2 8.593e-2
-2.143e-7 5.059e-6 5.697e-5 4.005e-4 1.963e-3 7.108e-3 1.951e-2 4.087e-2 6.457e-2 7.475e-2 5.774e-2 2.167e-2
-9.135e-7 1.958e-5 1.979e-4 1.231e-3 5.243e-3 1.610e-2 3.623e-2 5.900e-2 6.621e-2 4.494e-2 1.080e-2 1.697e-3
-2.914e-6 5.680e-5 5.162e-4 2.842e-3 1.049e-2 2.710e-2 4.904e-2 5.936e-2 4.172e-2 9.656e-3 2.093e-3 2.801e-2
-7.663e-6 1.361e-4 1.113e-3 5.419e-3 1.726e-2 3.713e-2 5.256e-2 4.372e-2 1.390e-2 4.503e-4 2.222e-2 3.847e-2
-1.733e-5 2.811e-4 2.071e-3 8.921e-3 2.445e-2 4.323e-2 4.603e-2 2.261e-2 5.668e-4 1.264e-2 3.386e-2 1.985e-2
-3.498e-5 5.198e-4 3.458e-3 1.317e-2 3.092e-2 4.420e-2 3.326e-2 6.546e-3 3.409e-3 2.597e-2 2.552e-2 2.222e-3
-6.299e-5 8.594e-4 5.173e-3 1.744e-2 3.488e-2 3.942e-2 1.881e-2 1.396e-4 1.355e-2 2.764e-2 9.543e-3 2.026e-3
-1.008e-4 1.269e-3 6.938e-3 2.074e-2 3.527e-2 3.072e-2 7.566e-3 2.089e-3 2.092e-2 1.906e-2 6.325e-4 1.153e-2
