--- data/franck-condon_legacy/nicholls_morse_fc.csv
@@ -1,2 +1,0 @@
-3.3636e-9 9.7126e-8 1.3505e-6 1.2033e-5 7.7140e-5 3.7873e-4 1.4800e-3 4.7231e-3 1.2534e-2 2.8020e-2 5.3261e-2
-3.9676e-8 1.0381e-6 1.2945e-5 1.0219e-4 5.7192e-4 2.4074e-3 7.8855e-3 2.0493e-2 4.2622e-2 7.0841e-2 9.2709e-2
