--- data/molecular_masses.csv
@@ -1,3 +1,0 @@
-molecule,mass_1,mass_2
-o2,15.999,15.999
-o2+,15.999,15.999
