--- data/test.csv
@@ -1,18 +1,0 @@
-wavenumber,intensity,branch,triplet,rot_qn,predissociation
-30911.20870683394,0.4837852897200883,r,1,8,0.104
-30911.2825584565,0.9766170859297814,r,2,8,0.175
-30911.410174334058,0.4929113325184463,r,3,8,0.158
-30911.429529464185,0.22950038923942714,p,1,2,0.122
-30911.7867512885,0.46341623464646015,p,2,2,0.163
-30912.130934027704,0.10445063262990838,r,3,18,0.19
-30912.204568712186,0.2339857619813228,p,3,2,0.121
-30912.5003394165,0.20705596168389157,r,2,18,0.212
-30912.919162178543,0.1026183753591225,r,1,18,0.097
-30917.013635470394,0.15506441191882286,p,3,12,0.182
-30917.1068238485,0.30733172018419713,p,2,12,0.194
-30917.25000691741,0.1522877859015359,p,1,12,0.101
-30918.71952201494,0.4217542292527649,r,1,6,0.121
-30918.748334884815,0.06260118515690749,p,1,0,0.13
-30918.8955632525,0.8514950962423588,r,2,6,0.188
-30919.130402752726,0.42982404674025937,r,3,6,0.165
-30919.2145776685,0.12643839125936482,p,2,0,0.001
