--- docs/todo.md
@@ -1,82 +1,0 @@
-# todo
-
-## General Theory
-
-### Molecular Approximations
-
-- Rigid Rotator
-    - Theory
-    - Energy levels
-    - Spectrum
-- Harmonic Oscillator
-- Anharmonic Oscillator
-- Nonrigid Rotator
-- Vibrating Rotator
-- Symmetric Top
-
-### Structure of Electronic Transitions
-
-- Total
-- Electronic
-- Vibrational
-- Rotational
-    - Branches
-
-### Intensities
-
-- Vibrational
-    - Partition function
-    - Intensity formulae
-    - Franck-Condon factors
-- Rotational
-    - Partition function
-    - Intensity formulae
-    - Hönl-London factors
-
-### Quantum Numbers
-
-- Atoms
-    - Single electron
-    - Multi electron
-    - Term symbols
-- Molecules
-    - Single electron
-    - Multi electron
-    - Term symbols
-
-## Detailed Theory
-
-### Hund's Coupling Cases
-
-- Case (a)
-    - Good quantum numbers
-    - Selection rules
-    - Term values
-- Case (b)
-    - Good quantum numbers
-    - Selection rules
-    - Term values
-
-### Uncoupling
-
-- Λ Doubling
-- Spin uncoupling
-- Term values
-
-### Electronic Transitions
-
-- Σ-Σ
-    - Branches
-- Σ-Π
-- Π-Σ
-- Π-Π
-
-## Individual Treatment
-
-### O2
-
-### O2+
-
-### NO
-
-### CO
